fix: Use onInput instead of onChange
Due to the fact that the toolkit React components are built with web components under the hood, it means that input components like VSCodeTextField or VSCodeTextArea follow the native browser change and input event model. This means the onChange event is fired when focus is blurred away from the input element and onInput is fired on every keystroke.